[Decomp][ObjectMgr] Cohesion split: extract gameobject spawn loaders to ObjectMgrGameObjects.cpp
