+++ src/game/Object/ObjectMgr.cpp
@@ -1568,201 +1568,8 @@ void ObjectMgr::RemoveCreatureFromGrid(uint32 guid, CreatureData const* data)
     }
 }
 
-/**
- * @brief Loads gameobject spawn records and validates their database data.
- */
-void ObjectMgr::LoadGameObjects()
-{
-    uint32 count = 0;
-
-    //                                                           0                1              2               3                      4                      5                      6
-    QueryResult* result = WorldDatabase.Query("SELECT `gameobject`.`guid`, `gameobject`.`id`, `gameobject`.`map`, `gameobject`.`position_x`, `gameobject`.`position_y`, `gameobject`.`position_z`, `gameobject`.`orientation`, "
-                          //             7                         8                         9                         10                        11                            12                           13                    14
-                          "`gameobject`.`rotation0`, `gameobject`.`rotation1`, `gameobject`.`rotation2`, `gameobject`.`rotation3`, `gameobject`.`spawntimesecs`, `gameobject`.`animprogress`, `gameobject`.`state`, `gameobject`.`spawnMask`,"
-                          //             15                      16                          17
-                          "`game_event_gameobject`.`event`, `pool_gameobject`.`pool_entry`, `pool_gameobject_template`.`pool_entry` "
-                          "FROM `gameobject` "
-                          "LEFT OUTER JOIN `game_event_gameobject` ON `gameobject`.`guid` = `game_event_gameobject`.`guid` "
-                          "LEFT OUTER JOIN `pool_gameobject` ON `gameobject`.`guid` = `pool_gameobject`.`guid` "
-                          "LEFT OUTER JOIN `pool_gameobject_template` ON `gameobject`.`id` = `pool_gameobject_template`.`id`");
-
-    if (!result)
-    {
-        BarGoLink bar(1);
-        bar.step();
-        sLog.outErrorDb(">> Loaded 0 gameobjects. DB table `gameobject` is empty.");
-        sLog.outString();
-        return;
-    }
-
-    BarGoLink bar(result->GetRowCount());
-
-    do
-    {
-        Field* fields = result->Fetch();
-        bar.step();
-
-        uint32 guid         = fields[ 0].GetUInt32();
-        uint32 entry        = fields[ 1].GetUInt32();
-
-        if (DisableMgr::IsDisabledFor(DISABLE_TYPE_GAMEOBJECT_SPAWN, entry, NULL, 0, guid))
-        {
-            sLog.outDebug("Gameobject guid %u (entry %u) spawning is disabled.", guid, entry);
-            continue;
-        }
-
-        GameObjectInfo const* gInfo = GetGameObjectInfo(entry);
-        if (!gInfo)
-        {
-            sLog.outErrorDb("Table `gameobject` has gameobject (GUID: %u) with non existing gameobject entry %u, skipped.", guid, entry);
-            continue;
-        }
-
-        if (gInfo->displayId && !sGameObjectDisplayInfoStore.LookupEntry(gInfo->displayId))
-        {
-            sLog.outErrorDb("Gameobject (GUID: %u Entry %u GoType: %u) have invalid displayId (%u), not loaded.", guid, entry, gInfo->type, gInfo->displayId);
-            continue;
-        }
-
-        GameObjectData& data = mGameObjectDataMap[guid];
-
-        data.id             = entry;
-        data.mapid          = fields[ 2].GetUInt32();
-        data.posX           = fields[ 3].GetFloat();
-        data.posY           = fields[ 4].GetFloat();
-        data.posZ           = fields[ 5].GetFloat();
-        data.orientation    = fields[ 6].GetFloat();
-        data.rotation0      = fields[ 7].GetFloat();
-        data.rotation1      = fields[ 8].GetFloat();
-        data.rotation2      = fields[ 9].GetFloat();
-        data.rotation3      = fields[10].GetFloat();
-        data.spawntimesecs  = fields[11].GetInt32();
-        data.animprogress   = fields[12].GetUInt32();
-        uint32 go_state     = fields[13].GetUInt32();
-        data.spawnMask      = fields[14].GetUInt8();
-        int16 gameEvent     = fields[15].GetInt16();
-        int16 GuidPoolId    = fields[16].GetInt16();
-        int16 EntryPoolId   = fields[17].GetInt16();
-
-        MapEntry const* mapEntry = sMapStore.LookupEntry(data.mapid);
-        if (!mapEntry)
-        {
-            sLog.outErrorDb("Table `gameobject` have gameobject (GUID: %u Entry: %u) that spawned at nonexistent map (Id: %u), skip", guid, data.id, data.mapid);
-            continue;
-        }
-
-        if (mapEntry->IsDungeon())
-        {
-            if (data.spawnMask & ~SPAWNMASK_DUNGEON_ALL)
-                sLog.outErrorDb("Table `gameobject` have gameobject (GUID: %u Entry: %u) that have wrong spawn mask %u for dungeon map (Id: %u), skip", guid, data.id, data.spawnMask, data.mapid);
-        }
-        else
-        {
-            if (data.spawnMask & ~SPAWNMASK_REGULAR)
-                sLog.outErrorDb("Table `gameobject` have gameobject (GUID: %u Entry: %u) that have wrong spawn mask %u for non-dungeon map (Id: %u), skip", guid, data.id, data.spawnMask, data.mapid);
-        }
-
-        if (data.spawntimesecs == 0 && gInfo->IsDespawnAtAction())
-        {
-            sLog.outErrorDb("Table `gameobject` have gameobject (GUID: %u Entry: %u) with `spawntimesecs` (0) value, but gameobejct marked as despawnable at action.", guid, data.id);
-        }
-
-        if (go_state >= MAX_GO_STATE)
-        {
-            sLog.outErrorDb("Table `gameobject` have gameobject (GUID: %u Entry: %u) with invalid `state` (%u) value, skip", guid, data.id, go_state);
-            continue;
-        }
-        data.go_state       = GOState(go_state);
-
-        if (data.rotation0 < -1.0f || data.rotation0 > 1.0f)
-        {
-            sLog.outErrorDb("Table `gameobject` have gameobject (GUID: %u Entry: %u) with invalid rotation0 (%f) value, skip", guid, data.id, data.rotation0);
-            continue;
-        }
-
-        if (data.rotation1 < -1.0f || data.rotation1 > 1.0f)
-        {
-            sLog.outErrorDb("Table `gameobject` have gameobject (GUID: %u Entry: %u) with invalid rotation1 (%f) value, skip", guid, data.id, data.rotation1);
-            continue;
-        }
-
-        if (data.rotation2 < -1.0f || data.rotation2 > 1.0f)
-        {
-            sLog.outErrorDb("Table `gameobject` have gameobject (GUID: %u Entry: %u) with invalid rotation2 (%f) value, skip", guid, data.id, data.rotation2);
-            continue;
-        }
-
-        if (data.rotation3 < -1.0f || data.rotation3 > 1.0f)
-        {
-            sLog.outErrorDb("Table `gameobject` have gameobject (GUID: %u Entry: %u) with invalid rotation3 (%f) value, skip", guid, data.id, data.rotation3);
-            continue;
-        }
-
-        if (!MapManager::IsValidMapCoord(data.mapid, data.posX, data.posY, data.posZ, data.orientation))
-        {
-            sLog.outErrorDb("Table `gameobject` have gameobject (GUID: %u Entry: %u) with invalid coordinates, skip", guid, data.id);
-            continue;
-        }
-
-        if (gameEvent == 0 && GuidPoolId == 0 && EntryPoolId == 0) // if not this is to be managed by GameEvent System or Pool system
-        {
-            AddGameobjectToGrid(guid, &data);
-        }
 
 
-        ++count;
-    }
-    while (result->NextRow());
-
-    delete result;
-
-    sLog.outString();
-    sLog.outString(">> Loaded %zu gameobjects", mGameObjectDataMap.size());
-}
-
-/**
- * @brief Adds a gameobject spawn GUID to the grid lookup for its map cell.
- *
- * @param guid The gameobject spawn GUID.
- * @param data The gameobject spawn data.
- */
-void ObjectMgr::AddGameobjectToGrid(uint32 guid, GameObjectData const* data)
-{
-    uint8 mask = data->spawnMask;
-    for (uint8 i = 0; mask != 0; ++i, mask >>= 1)
-    {
-        if (mask & 1)
-        {
-            CellPair cell_pair = MaNGOS::ComputeCellPair(data->posX, data->posY);
-            uint32 cell_id = (cell_pair.y_coord * TOTAL_NUMBER_OF_CELLS_PER_MAP) + cell_pair.x_coord;
-
-            CellObjectGuids& cell_guids = mMapObjectGuids[MAKE_PAIR32(data->mapid, i)][cell_id];
-            cell_guids.gameobjects.insert(guid);
-        }
-    }
-}
-
-/**
- * @brief Removes a gameobject spawn GUID from the grid lookup for its map cell.
- *
- * @param guid The gameobject spawn GUID.
- * @param data The gameobject spawn data.
- */
-void ObjectMgr::RemoveGameobjectFromGrid(uint32 guid, GameObjectData const* data)
-{
-    uint8 mask = data->spawnMask;
-    for (uint8 i = 0; mask != 0; ++i, mask >>= 1)
-    {
-        if (mask & 1)
-        {
-            CellPair cell_pair = MaNGOS::ComputeCellPair(data->posX, data->posY);
-            uint32 cell_id = (cell_pair.y_coord * TOTAL_NUMBER_OF_CELLS_PER_MAP) + cell_pair.x_coord;
-
-            CellObjectGuids& cell_guids = mMapObjectGuids[MAKE_PAIR32(data->mapid, i)][cell_id];
-            cell_guids.gameobjects.erase(guid);
-        }
-    }
-}
 
 // name must be checked to correctness (if received) before call this function
 ObjectGuid ObjectMgr::GetPlayerGuidByName(std::string name) const
